// src/app/features/stock-page/enums/stock.paths.enum.ts
export enum STOCK_API_PATHS {
     LIST = 'stock',
     DETAILS = 'stock/details/',
     POPULAR_ASSORTS = 'stock/popular/assorts',
     POPULAR_FABRICS = 'stock/popular/fabrics',
     POPULAR_CLOTHES = 'stock/popular/clothes',
     LACK_IN_STOCK = 'stock/lack',
     PATTERNS = 'stock/patterns',
     OPTIONS_LIST = 'stock/options/list',
     STATUS = 'stock/status/',
     // SAVE_SHOES_ORDER = 'stock/save/shoes',
     // SAVE_FABRIC_ORDER = 'stock/save/fabric',
     // SAVE_ASSORT_ORDER = 'stock/save/assort',

     // FABRIC_ADD = 'stock/add/fabric',
     // FABRIC_ORDER = 'stock/order/fabric',
     // SHOES_ORDER = 'stock/order/shoes',
     // ASSORT_ORDER = 'stock/order/assort',

     FABRIC_ADD = 'stock/fabric/add',
     FABRIC_ORDER = 'stock/fabric/order',
     FABRIC_UPDATE = 'stock/fabric/update/',
     SHOES_ORDER = 'stock/shoes/order',
     SHOES_UPDATE = 'stock/shoes/update/',
     ASSORT_ORDER = 'stock/assort/order',
     ASSORT_UPDATE = 'stock/assort/update/',
     MOVE_TO_STOCK = 'stock/fabrick/tostock/',
}

export enum STOCK_API_OPTIONS {
     WAREHOUSE = 'warehouse',
     ASSORT = 'assort',
     STATUS = 'status',
}